requirements.txt
appdirs==1.4.3
asgiref==3.3.1
beautifulsoup4==4.9.3
behave==1.2.6
behave-django==1.4.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Faker==6.6.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
parse==1.19.0
parse-type==0.5.2
pep517==0.8.2
progress==1.5
psycopg2==2.8.6
PyJWT==2.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
selenium==3.141.0
sentry-sdk==1.0.0
six==1.14.0
soupsieve==2.2
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.25.8
webencodings==0.5.1
